requirements.txt
streamlit
langchain
chromadb
faiss-cpu
openai
python-dotenv
